requirements.txt
sqlalchemy==2.0.29
pymysql==1.1.0
pandas==2.2.1
python-dotenv==1.0.1
uuid==1.30